refactor: move Inspectors to Inspection.
